Share task with others by email and Category CRUD operations

+++ src/main/java/com/personalproject/doit/services/TaskService.java
@@ -10,6 +10,7 @@ import com.personalproject.doit.exceptions.DatabaseException;
 import com.personalproject.doit.exceptions.ResourceNotFoundException;
 import com.personalproject.doit.repositories.CategoryRepository;
 import com.personalproject.doit.repositories.TaskRepository;
+import com.personalproject.doit.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
@@ -24,11 +25,13 @@ public class TaskService {
 
     private TaskRepository taskRepository;
     private CategoryRepository categoryRepository;
+    private UserRepository userRepository;
 
     @Autowired
-    public TaskService(TaskRepository taskRepository, CategoryRepository categoryRepository) {
+    public TaskService(TaskRepository taskRepository, CategoryRepository categoryRepository, UserRepository userRepository) {
         this.taskRepository = taskRepository;
         this.categoryRepository = categoryRepository;
+        this.userRepository = userRepository;
     }
 
     @Transactional(readOnly = true)
@@ -87,7 +90,22 @@ public class TaskService {
 
     @Transactional
     public void removeUserFromTask(Long id, Long userId) {
+        if (!taskRepository.existsById(id) && !userRepository.existsById(userId)) {
+            throw new ResourceNotFoundException("User or task id doesn't exist");
+        }
+
         taskRepository.removeUserFromTask(id, userId);
+    }
+
+    @Transactional
+    public void shareTask(Long id, String userEmail) {
+        Long userId = userRepository.getUserIdByEmail(userEmail);
+
+        if (userId == null) {
+            throw new ResourceNotFoundException("The user you want to share the task doesn't exist");
+        }
+
+        taskRepository.shareTask(userId, id);
     }
 
     private void copyDtoToEntity(TaskCategoryDTO dto, Task entity) {
